+++ src/Http/Requests/CreateShippingMethod.php
@@ -31,6 +31,7 @@ class CreateShippingMethod extends FormRequest implements CreateShippingMethodCo
             'name' => 'required|min:2|max:255',
             'carrier_id' => 'required|exists:carriers,id',
             'zone_id' => 'sometimes|nullable|exists:zones,id',
+            'tax_category_id' => 'sometimes|nullable|exists:tax_categories,id',
             'calculator' => ['sometimes', 'nullable', Rule::in(ShippingFeeCalculators::ids())],
             'configuration' => 'sometimes|json',
             'is_active' => 'sometimes|boolean',